requirements.txt
gensim==4.3.2
hdbscan
matplotlib==3.8.2
nltk==3.8.1
pandas==2.1.4
plotly==5.18.0
plotly-express==0.4.1
pyLDAvis==3.4.1
scikit-learn==1.4.0
streamlit==1.30.0
umap-learn==0.5.5
wordcloud==1.9.3